+++ src/shape_inference.hpp
@@ -1,0 +1,14 @@
+/*
+ * SPDX-FileCopyrightText: Copyright 2026 Arm Limited and/or its affiliates <open-source-office@arm.com>
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#pragma once
+
+namespace mlsdk::scenariorunner {
+
+struct ScenarioSpec;
+
+void resolveScenarioShapes(ScenarioSpec &scenarioSpec);
+
+} // namespace mlsdk::scenariorunner